Manage notifications better
Server goes crazy when we try to dismiss notifications on every navigation. As of now, server only supports dismissing all notifications. But this is annoying for users as notifications don't really go away. So we need to handle this in better way which doesn't put load on server. We save the dismissed notifications in `AsyncStorage`, and avoid showing them. When there are no notifications left, we dismiss all notifications.